diagonals returns diagonals given a board, possible wins includes diagonals
Rewrote the possible wins test to just check to see if each value in
the expected results collection is contained in the result of calling
the function possible wins.

Since we are testing for behavior and not implementation we don’t care
about the order of the result of possible wins, we just care that it
contains digonals, rows, and columns.