Players info returns the player information for the game
Players info now uses the result of calling get player one info, then
constructs that info along with getting the input with players 2 info
into a vector of maps.

We deleted get-player-two info as a result.

Line 55 since we changed what type of input we expect for
ask-player-for-move we has to pass in the right input in this line.

Ask player for turn order we also changed what type of input we expect
and changed test and implementation accordingly.