get move for a player takes a message and a board instead of a name and a board
Reasoning for this is to get-move for computer can work in the same
way, by passing it a message and a board.

Get game mode got moved to the input namespace with the other similar
functions.

Ask player for turn order message now takes a name that is a string to
construct the message

Changed negamax get move name to best move because of the function
names colliding with each other.